requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
notion-client==2.1.0
pyyaml>=6.0.1
python-frontmatter==1.0.1
jinja2==3.1.2
